Move MBLayer and children to their own package
